+++ NEWS.md
@@ -1,4 +1,23 @@
-## **_0.5-3_**
+
+## [zen4R 0.6](https://github.com/eblondel/zen4R) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-unavailable-red.svg)](https://github.com/eblondel/zen4R)
+
+**New features**
+
+* [#66](https://github.com/eblondel/zen4R/issues/66) Add record custom print method supporting various formats (internal + Zenodo supported export formats)
+* [#71](https://github.com/eblondel/zen4R/issues/71) Add `export_zenodo` function 
+* [#73](https://github.com/eblondel/zen4R/issues/73) Add mapping from published ZenodoRecord to [atom4R](https://github.com/eblondel/atom4R) DCEntry
+* [#74](https://github.com/eblondel/zen4R/issues/74) Package github-pages with pkgdown
+
+**Enhancements**
+
+* [#72](https://github.com/eblondel/zen4R/issues/72) Zenodo has new path for GET request --> avoid permanent redirect done by Zenodo
+
+**Bug fixes**
+
+* [#70](https://github.com/eblondel/zen4R/issues/70) Fix setEmbargoDate record method
+* [#74](https://github.com/eblondel/zen4R/issues/75) getRecords ElasticSearch query returns error if it includes spaces
+
+## [zen4R 0.5-3](https://cran.r-project.org/package=zen4R) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/package=zen4R)
 
 **Enhancements**
 
@@ -7,7 +26,7 @@
 * [#68](https://github.com/eblondel/zen4R/issues/68) Standardize R6-class R documentation
 
 
-## **_0.5-2_**
+## [zen4R 0.5-2](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.5-2.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.5-2.tar.gz)
 
 **Bug fixes**
 
@@ -15,7 +34,7 @@
 * [#63](https://github.com/eblondel/zen4R/issues/63) Fix new file upload API - case with filename with spaces
 * [#64](https://github.com/eblondel/zen4R/issues/64) Fix new file upload API - actual file upload
 
-## **_0.5-1_**
+## [zen4R 0.5-1](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.5-1.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.5-1.tar.gz)
 
 **Bug fixes**
 
@@ -23,7 +42,7 @@
 * [#58](https://github.com/eblondel/zen4R/pull/58) Fix getDepositionByConceptDOI method
 * [#60](https://github.com/eblondel/zen4R/issues/60) Fix old file upload API
 
-## **_0.5_**
+## [zen4R 0.5](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.5.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.5.tar.gz)
 
 **New features**
 
@@ -49,13 +68,13 @@
 * [#48](https://github.com/eblondel/zen4R/issues/48) Document download features in wiki + refine package description
 * [#49](https://github.com/eblondel/zen4R/issues/49) Add information on installation for Linux/OSX to documentation
 
-## **_0.4-1_**
+## [zen4R 0.4-1](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.4-1.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.4-1.tar.gz)
 
 **Improvements**
 
 *[#41](https://github.com/eblondel/zen4R/issues/41) Make ZenodoManager raise an error if token is invalid
 
-## **_0.4_**
+## [zen4R 0.4](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.4.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.4.tar.gz)
 
 **New features**
 
@@ -77,7 +96,7 @@
 * [#38](https://github.com/eblondel/zen4R/issues/38) download_zenodo() issue with md5sum
 * [#39](https://github.com/eblondel/zen4R/issues/39) getDepositionByConceptDOI / getRecordByConceptDOI doesn't return latest record
 
-## **_0.3_**
+## [zen4R 0.3](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.3.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.3.tar.gz)
 
 **New features**
 
@@ -87,7 +106,7 @@
 
 * [#24](https://github.com/eblondel/zen4R/issues/24) Exact search/matching with concept DOI doesn't work
 
-## **_0.2_**
+## [zen4R 0.2](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.2.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.2.tar.gz)
 
 **New features**
 
@@ -120,6 +139,6 @@
 
 _No issues_
 
-## **_0.1_**
+## [zen4R 0.1](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.1.tar.gz) | [![CRAN_Status_Badge](https://img.shields.io/badge/CRAN-published-blue.svg)](https://cran.r-project.org/src/contrib/Archive/zen4R/zen4R_0.1.tar.gz)
 
 First version